Update README with words and `.gitignore` PyCharm detritus
Does what it says on the tin.

The whole PR flow is maybe overkill for working with myself, but uh, I'm used to it and want to play around with github's PR system.

I'm not sure if I'll ever really get anywhere with this project because I really don't like writing code that much outside of work.